Add dark mode switcher and header with search
