+++ src/styles/App.scss
@@ -32,6 +32,7 @@ $primary: #fc58a7;
 }
 
 @import 'content';
+@import 'header';
 @import 'sidebar';
 @import 'transition';
 @import 'list';
