+++ src/styles/content.scss
@@ -13,13 +13,4 @@
   @include md {
     margin-left: 8rem;
   }
-
-  .header {
-    font-size: 3rem;
-    margin-left: 2rem;
-    @include sm {
-      margin-left: 1rem;
-      margin-top: 0rem;
-    }
-  }
 }
