+++ src/styles/dark.scss
@@ -1,4 +1,81 @@
+$back: #0f111a;
+$light-dark: #171a27;
 html.dark {
-  color: $light;
-  background-color: #0f111a;
+  background-color: $back;
+
+  .sidebar {
+    .logo .drawer {
+      &:hover {
+        background-color: $light-dark;
+      }
+    }
+    .menu {
+      a {
+        color: $gray;
+        &.active {
+          background: $light-dark;
+        }
+
+        &:hover {
+          background: linear-gradient(
+            90deg,
+            $primary,
+            lighten($light-dark, 20%)
+          );
+          transform: scale(1.05);
+          box-shadow: 0px 0px 20px rgba($primary, 0.1);
+          color: $back;
+        }
+
+        &:active {
+          transform: scale(0.95);
+        }
+      }
+    }
+  }
+
+  .header {
+    #search {
+      background-color: $light-dark;
+    }
+
+    .toolbar {
+      .search-box:focus-within {
+        #search {
+          background: $light-dark;
+        }
+      }
+      #search {
+        &:hover {
+          background: lighten($light-dark, 5%);
+        }
+      }
+    }
+
+    .change-theme {
+      .icon-theme {
+        background-image: url('/sun.svg');
+        background-color: $light-dark;
+        &:active {
+          background-color: rgba($color: $primary, $alpha: 0.5);
+        }
+
+        &:hover {
+          background-color: lighten($color: $light-dark, $amount: 5%);
+        }
+      }
+    }
+  }
+
+  .list {
+    .title {
+      color: $gray;
+    }
+    .array .card {
+      color: $dark;
+      h1 {
+        color: $gray;
+      }
+    }
+  }
 }
