+++ src/styles/header.scss
@@ -1,0 +1,71 @@
+.header {
+  margin-left: 2rem;
+  margin-bottom: 1.5rem;
+  @include sm {
+    margin-left: 1rem;
+    margin-top: 0rem;
+  }
+  h1 {
+    font-size: 3rem;
+    margin-bottom: 1.5rem;
+  }
+
+  .toolbar {
+    display: flex;
+    align-items: center;
+    margin-left: 1rem;
+    svg {
+      position: relative;
+      z-index: 3;
+      top: 3px;
+    }
+    .search-box:focus-within {
+      svg {
+        color: $primary;
+      }
+      #search {
+        width: 20rem;
+      }
+    }
+    #search {
+      margin-left: -2.5rem;
+      font-family: inherit;
+      font-size: 24px;
+      padding: 0.5rem 0.5rem 0.5rem 3rem;
+      border: none;
+      color: $gray;
+      background: $light;
+      border-radius: 12px;
+      width: 15rem;
+      transition: all 0.25s ease-in-out;
+      &:hover {
+        background-color: darken($light, 3%);
+      }
+    }
+  }
+
+  .change-theme {
+    display: inline-flex;
+    justify-content: center;
+    margin-left: 1rem;
+    .icon-theme {
+      background-color: $light;
+      &:active {
+        background-color: rgba($color: $primary, $alpha: 0.5);
+        transform: scale(1.5);
+      }
+      &:hover {
+        background-color: darken($color: $light, $amount: 5%);
+      }
+      border-radius: 12px;
+      transition: all 0.1s ease-in-out;
+      width: 24px;
+      height: 24px;
+      padding: 11px;
+      background-size: 24px 24px;
+      background-position: center;
+      background-repeat: no-repeat;
+      background-image: url('/moon.svg');
+    }
+  }
+}
