+++ src/views/home.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import List from '../components/list.js';
+import Header from '../components/header';
 
 export default function Home(props: Object) {
   return (
     <div className='home'>
-      <h1 className='header'>Home</h1>
+      <Header title='Home' />
+
       <List title="KGP's picks 🔥" redirect />
       <List title='Top Weekly 🏅' redirect />
       <List title='Most Played Songs 🎵' redirect />
